change some uint64_t to size_t

+++ tokenizer_generator/include/finite_automaton/array/node_array.h
@@ -16,11 +16,11 @@ void kev_nodearray_destroy(KevNodeArray* array);
 bool kev_nodearray_push_back(KevNodeArray* array, KevGraphNode* node);
 static inline void kev_nodearray_pop_back(KevNodeArray* array);
 bool kev_nodearray_expand(KevNodeArray* array);
-static inline void kev_nodearray_assign(KevNodeArray* array, uint64_t index, KevGraphNode* value);
-static inline KevGraphNode* kev_nodearray_visit(KevNodeArray* array, uint64_t index);
-static inline uint64_t kev_nodearray_size(KevNodeArray* array);
+static inline void kev_nodearray_assign(KevNodeArray* array, size_t index, KevGraphNode* value);
+static inline KevGraphNode* kev_nodearray_visit(KevNodeArray* array, size_t index);
+static inline size_t kev_nodearray_size(KevNodeArray* array);
 
-static inline uint64_t kev_nodearray_size(KevNodeArray* array) {
+static inline size_t kev_nodearray_size(KevNodeArray* array) {
   return array->current - array->begin;
 }
 
@@ -28,11 +28,11 @@ static inline void kev_nodearray_pop_back(KevNodeArray* array) {
   array->current--;
 }
 
-static inline KevGraphNode* kev_nodearray_visit(KevNodeArray* array, uint64_t index) {
+static inline KevGraphNode* kev_nodearray_visit(KevNodeArray* array, size_t index) {
   return array->begin[index];
 }
 
-static inline void kev_nodearray_assign(KevNodeArray* array, uint64_t index, KevGraphNode* value) {
+static inline void kev_nodearray_assign(KevNodeArray* array, size_t index, KevGraphNode* value) {
   array->begin[index] = value;
 }
