+++ tokenizer_generator/include/finite_automaton/queue/int_queue.h
@@ -4,26 +4,26 @@
 #include "tokenizer_generator/include/general/global_def.h"
 
 typedef struct tagKevIntQueue {
-  uint64_t* array;
-  uint64_t capacity;
-  uint64_t head;
-  uint64_t tail;
+  size_t* array;
+  size_t capacity;
+  size_t head;
+  size_t tail;
 } KevIntQueue;
 
 bool kev_intqueue_init(KevIntQueue* queue);
 void kev_intqueue_destroy(KevIntQueue* queue);
 
 bool kev_intqueue_expand(KevIntQueue* queue);
-static inline bool kev_intqueue_insert(KevIntQueue* queue, uint64_t element);
-static inline uint64_t kev_intqueue_pop(KevIntQueue* queue);
+static inline bool kev_intqueue_insert(KevIntQueue* queue, size_t element);
+static inline size_t kev_intqueue_pop(KevIntQueue* queue);
 
 static inline bool kev_intqueue_empty(KevIntQueue* queue) {
   return queue->head == queue->tail;
 }
 
-static inline bool kev_intqueue_insert(KevIntQueue* queue, uint64_t element) {
+static inline bool kev_intqueue_insert(KevIntQueue* queue, size_t element) {
   if (!queue) return false;
-  uint64_t new_tail = (queue->tail + 1) & (queue->capacity - 1);
+  size_t new_tail = (queue->tail + 1) & (queue->capacity - 1);
   if (new_tail == queue->head) {
     if (!kev_intqueue_expand(queue))
       return false;
@@ -35,9 +35,9 @@ static inline bool kev_intqueue_insert(KevIntQueue* queue, uint64_t element) {
   return true;
 }
 
-static inline uint64_t kev_intqueue_pop(KevIntQueue* queue) {
+static inline size_t kev_intqueue_pop(KevIntQueue* queue) {
   if (!queue || queue->head == queue->tail) return -1;
-  uint64_t retval = queue->array[queue->head];
+  size_t retval = queue->array[queue->head];
   queue->head = (queue->head + 1) & (queue->capacity - 1);
   return retval;
 }
